Adjusted map chart rendering error
