+++ src/containers/auth-zero/Auth/auth0-variables.js
@@ -1,6 +1,6 @@
 export const AUTH_CONFIG = {
   domain: "ladnetwork.auth0.com",
   clientId: "XQh6ZLLicvNOdZUMBrCvmL2zyrgmE1fB",
-  callbackUrl: "https://ladnetwork.netlify.com/callback"
-  // callbackUrl: "http://localhost:3000/callback"
+  // callbackUrl: "https://ladnetwork.netlify.com/callback"
+  callbackUrl: "http://localhost:3000/callback"
 };
